refactor(hmsl): move common options for CLI
