+++ ggshield/cmd/hmsl/query.py
@@ -7,7 +7,7 @@ import click
 from requests import HTTPError
 
 from ggshield.cmd.common_options import add_common_options
-from ggshield.cmd.hmsl.fingerprint import input_arg
+from ggshield.cmd.hmsl.hmsl_common_options import input_arg
 from ggshield.core.config import Config
 from ggshield.core.errors import ParseError, UnexpectedError
 from ggshield.core.text_utils import display_info, pluralize
