+++ ggshield/hmsl/collection.py
@@ -1,0 +1,14 @@
+from dataclasses import dataclass
+from typing import Dict, Optional, Set
+
+
+@dataclass
+class PreparedSecrets:
+    payload: Set[str]
+    mapping: Dict[str, str]
+
+
+@dataclass
+class SecretWithKey:
+    key: Optional[str]
+    value: str
